Co-locate Huma handlers with route groups
Splitting route registration into Huma groups still left the typed handlers and input schemas in a single omnibus file, so the API contract was only half organized by domain. Move the Huma handlers, request types, response wrappers, and endpoint-specific enums into the same files as their route groups so each API area owns its registration and implementation together.

Keep only shared Huma plumbing, common path/query helper inputs, error conversion, schema naming, and SSE/write helpers in the shared route file. Use gopls symbols for the declaration inventory and gopls imports/format/check after the mechanical move so import cleanup and diagnostics come from the language server.

Validation: gopls check over all Huma route files; go test -count=1 ./internal/server; npm run generate:api (from frontend, no generated diff); go test -count=1 ./...; git diff --check; git diff --cached --check
