+++ internal/server/huma_routes_analytics.go
@@ -1,5 +1,14 @@
 package server
 
+import (
+	"context"
+	"net/http"
+	"time"
+
+	"go.kenn.io/agentsview/internal/db"
+	"go.kenn.io/agentsview/internal/timeutil"
+)
+
 func (s *Server) registerAnalyticsRoutes() {
 	group := newRouteGroup(s.api, "/api/v1/analytics", "Analytics")
 
@@ -15,15 +24,242 @@ func (s *Server) registerAnalyticsRoutes() {
 	get(s, group, "/signals", "Get signal analytics", s.humaAnalyticsSignals)
 }
 
-func (s *Server) registerTrendsRoutes() {
-	group := newRouteGroup(s.api, "/api/v1/trends", "Trends")
+type analyticsGranularity string
 
-	get(s, group, "/terms", "Get trend terms", s.humaTrendsTerms)
+const (
+	analyticsGranularityDay   analyticsGranularity = "day"
+	analyticsGranularityWeek  analyticsGranularity = "week"
+	analyticsGranularityMonth analyticsGranularity = "month"
+)
+
+type heatmapMetric string
+
+const (
+	heatmapMetricMessages     heatmapMetric = "messages"
+	heatmapMetricSessions     heatmapMetric = "sessions"
+	heatmapMetricOutputTokens heatmapMetric = "output_tokens"
+)
+
+type topSessionMetric string
+
+const (
+	topSessionMetricMessages     topSessionMetric = "messages"
+	topSessionMetricDuration     topSessionMetric = "duration"
+	topSessionMetricOutputTokens topSessionMetric = "output_tokens"
+)
+
+type AnalyticsFilterInput struct {
+	From             string           `query:"from" format:"date" doc:"Range start date"`
+	To               string           `query:"to" format:"date" doc:"Range end date"`
+	Timezone         string           `query:"timezone" doc:"IANA timezone name"`
+	Machine          string           `query:"machine" doc:"Filter by machine"`
+	Project          string           `query:"project" doc:"Filter by project"`
+	Agent            string           `query:"agent" doc:"Filter by agent"`
+	DayOfWeek        optionalIntParam `query:"dow" minimum:"0" maximum:"6" doc:"Day of week, Monday=0 through Sunday=6"`
+	Hour             optionalIntParam `query:"hour" minimum:"0" maximum:"23" doc:"Hour of day, 0 through 23"`
+	MinUserMessages  int              `query:"min_user_messages" minimum:"0" doc:"Minimum user message count"`
+	ActiveSince      string           `query:"active_since" format:"date-time" doc:"Filter sessions active since this RFC3339 timestamp"`
+	IncludeOneShot   bool             `query:"include_one_shot" doc:"Include one-shot sessions"`
+	IncludeAutomated bool             `query:"include_automated" doc:"Include automated sessions"`
+	Termination      string           `query:"termination" doc:"Filter by termination reason"`
 }
 
-func (s *Server) registerUsageRoutes() {
-	group := newRouteGroup(s.api, "/api/v1/usage", "Usage")
+type analyticsActivityInput struct {
+	AnalyticsFilterInput
+	Granularity analyticsGranularity `query:"granularity" enum:"day,week,month" default:"day" doc:"Time bucket granularity"`
+}
 
-	get(s, group, "/summary", "Get usage summary", s.humaUsageSummary)
-	get(s, group, "/top-sessions", "Get top usage sessions", s.humaUsageTopSessions)
+type analyticsHeatmapInput struct {
+	AnalyticsFilterInput
+	Metric heatmapMetric `query:"metric" enum:"messages,sessions,output_tokens" default:"messages" doc:"Heatmap metric"`
+}
+
+type analyticsTopSessionsInput struct {
+	AnalyticsFilterInput
+	Metric topSessionMetric `query:"metric" enum:"messages,duration,output_tokens" default:"messages" doc:"Ranking metric"`
+}
+
+func analyticsFilterFromInput(in AnalyticsFilterInput) (db.AnalyticsFilter, error) {
+	tz := in.Timezone
+	if tz == "" {
+		tz = "UTC"
+	}
+	if _, err := time.LoadLocation(tz); err != nil {
+		return db.AnalyticsFilter{}, apiError(http.StatusBadRequest, "invalid timezone: "+tz)
+	}
+	from, to := defaultDateRange(in.From, in.To)
+	if !timeutil.IsValidDate(from) || !timeutil.IsValidDate(to) {
+		return db.AnalyticsFilter{}, apiError(http.StatusBadRequest, "invalid date format: use YYYY-MM-DD")
+	}
+	if from > to {
+		return db.AnalyticsFilter{}, apiError(http.StatusBadRequest, "from must not be after to")
+	}
+	if in.ActiveSince != "" && !timeutil.IsValidTimestamp(in.ActiveSince) {
+		return db.AnalyticsFilter{}, apiError(http.StatusBadRequest, "invalid active_since: use RFC3339 timestamp")
+	}
+	return db.AnalyticsFilter{
+		From:             from,
+		To:               to,
+		Machine:          in.Machine,
+		Project:          in.Project,
+		Agent:            in.Agent,
+		Timezone:         tz,
+		DayOfWeek:        optionalIntValue(in.DayOfWeek),
+		Hour:             optionalIntValue(in.Hour),
+		MinUserMessages:  in.MinUserMessages,
+		ExcludeOneShot:   !in.IncludeOneShot,
+		ExcludeAutomated: !in.IncludeAutomated,
+		ActiveSince:      in.ActiveSince,
+		Termination:      in.Termination,
+	}, nil
+}
+
+func (s *Server) humaAnalyticsSummary(
+	ctx context.Context,
+	in *AnalyticsFilterInput,
+) (*jsonOutput[db.AnalyticsSummary], error) {
+	f, err := analyticsFilterFromInput(*in)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsSummary(ctx, f)
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.AnalyticsSummary]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsActivity(
+	ctx context.Context,
+	in *analyticsActivityInput,
+) (*jsonOutput[db.ActivityResponse], error) {
+	f, err := analyticsFilterFromInput(in.AnalyticsFilterInput)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsActivity(ctx, f, string(in.Granularity))
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.ActivityResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsHeatmap(
+	ctx context.Context,
+	in *analyticsHeatmapInput,
+) (*jsonOutput[db.HeatmapResponse], error) {
+	f, err := analyticsFilterFromInput(in.AnalyticsFilterInput)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsHeatmap(ctx, f, string(in.Metric))
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.HeatmapResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsProjects(
+	ctx context.Context,
+	in *AnalyticsFilterInput,
+) (*jsonOutput[db.ProjectsAnalyticsResponse], error) {
+	f, err := analyticsFilterFromInput(*in)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsProjects(ctx, f)
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.ProjectsAnalyticsResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsHourOfWeek(
+	ctx context.Context,
+	in *AnalyticsFilterInput,
+) (*jsonOutput[db.HourOfWeekResponse], error) {
+	f, err := analyticsFilterFromInput(*in)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsHourOfWeek(ctx, f)
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.HourOfWeekResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsSessionShape(
+	ctx context.Context,
+	in *AnalyticsFilterInput,
+) (*jsonOutput[db.SessionShapeResponse], error) {
+	f, err := analyticsFilterFromInput(*in)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsSessionShape(ctx, f)
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.SessionShapeResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsVelocity(
+	ctx context.Context,
+	in *AnalyticsFilterInput,
+) (*jsonOutput[db.VelocityResponse], error) {
+	f, err := analyticsFilterFromInput(*in)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsVelocity(ctx, f)
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.VelocityResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsTools(
+	ctx context.Context,
+	in *AnalyticsFilterInput,
+) (*jsonOutput[db.ToolsAnalyticsResponse], error) {
+	f, err := analyticsFilterFromInput(*in)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsTools(ctx, f)
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.ToolsAnalyticsResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsTopSessions(
+	ctx context.Context,
+	in *analyticsTopSessionsInput,
+) (*jsonOutput[db.TopSessionsResponse], error) {
+	f, err := analyticsFilterFromInput(in.AnalyticsFilterInput)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsTopSessions(ctx, f, string(in.Metric))
+	if err != nil {
+		return nil, internalError("analytics error", err)
+	}
+	return &jsonOutput[db.TopSessionsResponse]{Body: result}, nil
+}
+
+func (s *Server) humaAnalyticsSignals(
+	ctx context.Context,
+	in *AnalyticsFilterInput,
+) (*jsonOutput[db.SignalsAnalyticsResponse], error) {
+	f, err := analyticsFilterFromInput(*in)
+	if err != nil {
+		return nil, err
+	}
+	result, err := s.db.GetAnalyticsSignals(ctx, f)
+	if err != nil {
+		return nil, internalError("analytics signals error", err)
+	}
+	return &jsonOutput[db.SignalsAnalyticsResponse]{Body: result}, nil
 }
