+++ internal/server/huma_routes_search.go
@@ -1,6 +1,14 @@
 package server
 
-import "net/http"
+import (
+	"context"
+	"errors"
+	"net/http"
+	"strings"
+
+	"go.kenn.io/agentsview/internal/db"
+	"go.kenn.io/agentsview/internal/service"
+)
 
 func (s *Server) registerSearchRoutes() {
 	group := newRouteGroup(s.api, "/api/v1/search", "Search")
@@ -9,9 +17,135 @@ func (s *Server) registerSearchRoutes() {
 	get(s, group, "/content", "Search session content", s.humaSearchContent)
 }
 
-func (s *Server) registerSecretsRoutes() {
-	group := newRouteGroup(s.api, "/api/v1/secrets", "Secrets")
+type searchSort string
 
-	get(s, group, "", "List secret findings", s.humaListSecrets)
-	stream(s, group, http.MethodPost, "/scan", "Scan secrets", s.humaScanSecrets)
+const (
+	searchSortRelevance searchSort = "relevance"
+	searchSortRecency   searchSort = "recency"
+)
+
+type contentSearchMode string
+
+const (
+	contentSearchModeSubstring contentSearchMode = "substring"
+	contentSearchModeRegex     contentSearchMode = "regex"
+	contentSearchModeFTS       contentSearchMode = "fts"
+)
+
+type searchInput struct {
+	Query   string     `query:"q" required:"true" doc:"Search query"`
+	Project string     `query:"project" doc:"Filter by project"`
+	Sort    searchSort `query:"sort" enum:"relevance,recency" default:"relevance" doc:"Sort order"`
+	Limit   int        `query:"limit" minimum:"0" doc:"Maximum number of results"`
+	Cursor  int        `query:"cursor" minimum:"0" doc:"Pagination cursor"`
+}
+
+type contentSearchInput struct {
+	Pattern          string            `query:"pattern" required:"true" doc:"Pattern to search for"`
+	Mode             contentSearchMode `query:"mode" enum:"substring,regex,fts" doc:"Search mode"`
+	In               string            `query:"in" doc:"Comma-separated content sources"`
+	ExcludeSystem    bool              `query:"exclude_system" doc:"Exclude system messages"`
+	Reveal           bool              `query:"reveal" doc:"Return unredacted secret matches for localhost callers"`
+	Project          string            `query:"project" doc:"Filter by project"`
+	ExcludeProject   string            `query:"exclude_project" doc:"Exclude a project"`
+	Machine          string            `query:"machine" doc:"Filter by machine"`
+	Agent            string            `query:"agent" doc:"Filter by agent"`
+	Date             string            `query:"date" format:"date" doc:"Filter to a single YYYY-MM-DD date"`
+	DateFrom         string            `query:"date_from" format:"date" doc:"Filter start date"`
+	DateTo           string            `query:"date_to" format:"date" doc:"Filter end date"`
+	ActiveSince      string            `query:"active_since" format:"date-time" doc:"Filter sessions active since this RFC3339 timestamp"`
+	IncludeChildren  bool              `query:"include_children" doc:"Include child sessions"`
+	IncludeAutomated bool              `query:"include_automated" doc:"Include automated sessions"`
+	IncludeOneShot   bool              `query:"include_one_shot" doc:"Include one-shot sessions"`
+	Limit            int               `query:"limit" minimum:"0" doc:"Maximum number of results"`
+	Cursor           int               `query:"cursor" minimum:"0" doc:"Pagination cursor"`
+}
+
+func (s *Server) humaSearch(
+	ctx context.Context,
+	in *searchInput,
+) (*jsonOutput[searchResponse], error) {
+	query := strings.TrimSpace(in.Query)
+	if query == "" {
+		return nil, apiError(http.StatusBadRequest, "query required")
+	}
+	if !s.db.HasFTS() {
+		return nil, apiError(http.StatusNotImplemented, "search not available")
+	}
+	limit := clampLimit(in.Limit, db.DefaultSearchLimit, db.MaxSearchLimit)
+	page, err := s.db.Search(ctx, db.SearchFilter{
+		Query:   prepareFTSQuery(query),
+		Project: in.Project,
+		Sort:    string(in.Sort),
+		Cursor:  in.Cursor,
+		Limit:   limit,
+	})
+	if err != nil {
+		return nil, serverError(err)
+	}
+	results := page.Results
+	if results == nil {
+		results = []db.SearchResult{}
+	}
+	return &jsonOutput[searchResponse]{
+		Body: searchResponse{
+			Query:   query,
+			Results: results,
+			Count:   len(results),
+			Next:    page.NextCursor,
+		},
+	}, nil
+}
+
+func (s *Server) humaSearchContent(
+	ctx context.Context,
+	in *contentSearchInput,
+) (*jsonOutput[*service.ContentSearchResult], error) {
+	if in.Reveal && !isLocalhostContext(ctx) {
+		return nil, apiError(http.StatusForbidden, "reveal is only permitted from localhost")
+	}
+	var sources []string
+	if in.In != "" {
+		sources = strings.Split(in.In, ",")
+	}
+	if err := validateDateFilterValues(in.Date, in.DateFrom, in.DateTo, in.ActiveSince); err != nil {
+		return nil, err
+	}
+	res, err := s.sessions.SearchContent(ctx, service.ContentSearchRequest{
+		Pattern:          in.Pattern,
+		Mode:             string(in.Mode),
+		Sources:          sources,
+		ExcludeSystem:    in.ExcludeSystem,
+		Reveal:           in.Reveal,
+		Project:          in.Project,
+		ExcludeProject:   in.ExcludeProject,
+		Machine:          in.Machine,
+		Agent:            in.Agent,
+		Date:             in.Date,
+		DateFrom:         in.DateFrom,
+		DateTo:           in.DateTo,
+		ActiveSince:      in.ActiveSince,
+		IncludeChildren:  in.IncludeChildren,
+		IncludeAutomated: in.IncludeAutomated,
+		IncludeOneShot:   in.IncludeOneShot,
+		Limit:            in.Limit,
+		Cursor:           in.Cursor,
+	})
+	if err != nil {
+		if handled := handleHumaContextError(err); handled != nil {
+			return nil, handled
+		}
+		if handled := handleHumaReadOnly(err); handled != nil {
+			return nil, handled
+		}
+		var inputErr *db.SearchInputError
+		if errors.As(err, &inputErr) {
+			return nil, apiError(http.StatusBadRequest, err.Error())
+		}
+		return nil, apiError(http.StatusInternalServerError, err.Error())
+	}
+	if res.Matches == nil {
+		res.Matches = []db.ContentMatch{}
+	}
+	return &jsonOutput[*service.ContentSearchResult]{Body: res}, nil
 }
