+++ internal/server/huma_routes_sync.go
@@ -1,6 +1,16 @@
 package server
 
-import "net/http"
+import (
+	"context"
+	"errors"
+	"net/http"
+	"time"
+
+	"github.com/danielgtaylor/huma/v2"
+	"go.kenn.io/agentsview/internal/db"
+	"go.kenn.io/agentsview/internal/service"
+	syncpkg "go.kenn.io/agentsview/internal/sync"
+)
 
 func (s *Server) registerSyncRoutes() {
 	group := newRouteGroup(s.api, "/api/v1", "Sync")
@@ -9,4 +19,100 @@ func (s *Server) registerSyncRoutes() {
 	stream(s, group, http.MethodPost, "/resync", "Trigger full resync", s.humaTriggerResync)
 	get(s, group, "/sync/status", "Get sync status", s.humaSyncStatus)
 	post(s, group, "/sessions/sync", "Sync a session", s.humaSyncSession)
+}
+
+type syncStatusResponse struct {
+	LastSync string             `json:"last_sync"`
+	Stats    *syncpkg.SyncStats `json:"stats"`
+}
+
+type syncStatsResponse = syncpkg.SyncStats
+
+type sessionSyncInput struct {
+	Body service.SyncInput
+}
+
+func (s *Server) humaSyncStatus(
+	_ context.Context,
+	_ *emptyInput,
+) (*jsonOutput[syncStatusResponse], error) {
+	if s.engine == nil {
+		return &jsonOutput[syncStatusResponse]{Body: syncStatusResponse{}}, nil
+	}
+	lastSync := s.engine.LastSync()
+	stats := s.engine.LastSyncStats()
+	var lastSyncStr string
+	if !lastSync.IsZero() {
+		lastSyncStr = lastSync.Format(time.RFC3339)
+	}
+	return &jsonOutput[syncStatusResponse]{
+		Body: syncStatusResponse{LastSync: lastSyncStr, Stats: &stats},
+	}, nil
+}
+
+func (s *Server) humaTriggerSync(
+	ctx context.Context,
+	_ *emptyInput,
+) (*huma.StreamResponse, error) {
+	if s.engine == nil {
+		return nil, apiError(http.StatusNotImplemented, "not available in remote mode")
+	}
+	return &huma.StreamResponse{Body: func(hctx huma.Context) {
+		stream, ok := newHumaSSEStream(hctx)
+		if !ok {
+			stats := s.engine.SyncAll(ctx, nil)
+			writeHumaJSON(hctx, http.StatusOK, stats)
+			return
+		}
+		stats := s.engine.SyncAll(ctx, func(p syncpkg.Progress) {
+			stream.SendJSON("progress", p)
+		})
+		stream.SendJSON("done", stats)
+	}}, nil
+}
+
+func (s *Server) humaTriggerResync(
+	ctx context.Context,
+	_ *emptyInput,
+) (*huma.StreamResponse, error) {
+	if s.engine == nil {
+		return nil, apiError(http.StatusNotImplemented, "not available in remote mode")
+	}
+	return &huma.StreamResponse{Body: func(hctx huma.Context) {
+		stream, ok := newHumaSSEStream(hctx)
+		if !ok {
+			stats := s.engine.ResyncAll(ctx, nil)
+			writeHumaJSON(hctx, http.StatusOK, stats)
+			return
+		}
+		stats := s.engine.ResyncAll(ctx, func(p syncpkg.Progress) {
+			stream.SendJSON("progress", p)
+		})
+		stream.SendJSON("done", stats)
+	}}, nil
+}
+
+func (s *Server) humaSyncSession(
+	ctx context.Context,
+	in *sessionSyncInput,
+) (*jsonOutput[*service.SessionDetail], error) {
+	if (in.Body.Path == "" && in.Body.ID == "") ||
+		(in.Body.Path != "" && in.Body.ID != "") {
+		return nil, apiError(http.StatusBadRequest, "exactly one of 'path' or 'id' is required")
+	}
+	detail, err := s.sessions.Sync(ctx, in.Body)
+	if err != nil {
+		if handled := handleHumaContextError(err); handled != nil {
+			return nil, handled
+		}
+		if handled := handleHumaReadOnly(err); handled != nil {
+			return nil, handled
+		}
+		if errors.Is(err, db.ErrSessionExcluded) ||
+			errors.Is(err, db.ErrSessionTrashed) {
+			return nil, apiError(http.StatusConflict, err.Error())
+		}
+		return nil, apiError(http.StatusInternalServerError, err.Error())
+	}
+	return &jsonOutput[*service.SessionDetail]{Body: detail}, nil
 }
